Refactor referencer imports and remove unused referencer module and constants
