+++ microSALT/utils/job_creator.py
@@ -18,7 +18,7 @@ from pathlib import Path
 import yaml
 
 from microSALT.store.db_manipulator import DB_Manipulator
-from microSALT.utils.referencer.referencer import Referencer
+from microSALT.utils.referencer import Referencer
 
 
 class Job_Creator:
