--- tests/referencer/test_utils.py
@@ -1,61 +1,0 @@
-import pytest
-from microSALT.utils.referencer.utils import (
-    get_species,
-    is_escherichia,
-    is_klebsiella,
-    get_reference_if_enterobacteriaceae,
-)
-from microSALT.utils.referencer.constants import (
-    ESCHERICHIA_REFERENCE,
-    KLEBSIELLA_REFERENCE,
-    Escherichia,
-    Klebsiella,
-)
-
-
-@pytest.mark.parametrize(
-    "organism_name, expected_species",
-    [
-        ("Escherichia_coli", "coli"),
-        ("Klebsiella_pneumoniae", "pneumoniae"),
-        ("Escherichia_coli_O157:H7", "o157:h7"),
-    ],
-)
-def test_get_species(organism_name, expected_species):
-    assert get_species(organism_name) == expected_species
-
-
-@pytest.mark.parametrize(
-    "species_name, expected_result",
-    [
-        (Escherichia.ALBA, True),
-        (Escherichia.COLI, True),
-        (Klebsiella.ELECTRICA, False),
-    ],
-)
-def test_is_escherichia(species_name, expected_result):
-    assert is_escherichia(species_name) == expected_result
-
-
-@pytest.mark.parametrize(
-    "species_name, expected_result",
-    [
-        (Klebsiella.PNEUMONIAE, True),
-        (Klebsiella.OXYTOCA, True),
-        (Escherichia.ALBA, False),
-    ],
-)
-def test_is_klebsiella(species_name, expected_result):
-    assert is_klebsiella(species_name) == expected_result
-
-
-@pytest.mark.parametrize(
-    "organism_name, expected_reference",
-    [
-        ("escherichia_coli", ESCHERICHIA_REFERENCE),
-        ("klebsiella_pneumoniae", KLEBSIELLA_REFERENCE),
-        ("other_species", "other_species"),
-    ],
-)
-def test_get_reference_if_enterobacteriaceae(organism_name, expected_reference):
-    assert get_reference_if_enterobacteriaceae(organism_name) == expected_reference
